requirements.txt
Babel==2.6.0
backcall==0.1.0
certifi==2018.8.13
chardet==3.0.4
decorator==4.3.0
Django==2.1
django-anymail==3.0
django-constance==2.2.0
django-environ==0.4.5
django-modeltranslation==0.13b1
django-phonenumber-field==2.0.0
django-picklefield==1.0.0
django-rosetta==0.8.3
idna==2.7
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
microsofttranslator==0.8
parso==0.3.1
pexpect==4.6.0
phonenumberslite==8.9.10
pickleshare==0.7.4
polib==1.1.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
